@@ -3,7 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@apollo/react-hooks": "^3.1.5",
+    "@apollo/client": "^3.3.6",
+    "@apollo/react-hooks": "^4.0.0",
     "@bugsnag/js": "^7.0.0",
     "@bugsnag/plugin-react": "^7.0.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.28",
@@ -21,14 +22,14 @@
     "@types/react-gravatar": "^2.6.7",
     "@types/react-transition-group": "^4.2.4",
     "@types/storybook__addon-links": "^3.3.4",
-    "apollo-boost": "^0.3.1",
-    "apollo-client": "^2.5.1",
+    "apollo-cache-inmemory": "^1.6.6",
+    "apollo-client": "^2.6.10",
     "apollo-link-context": "^1.0.17",
     "autoprefixer": "^9.7.4",
     "axios": "^0.18.0",
     "framer": "^1.2.2",
     "framer-motion": "^1.11.0",
-    "graphql": "^14.1.1",
+    "graphql": "^15.4.0",
     "moment": "^2.24.0",
     "node-sass": "^4.11.0",
     "postcss-cli": "^7.1.0",
@@ -39,9 +40,9 @@
     "react-dom": "^16.13.1",
     "react-gravatar": "^2.6.3",
     "react-markdown": "^4.0.6",
-    "react-scripts": "2.1.8",
+    "react-scripts": "4.0.1",
     "tailwindcss": "^1.2.0",
-    "typescript": "3.3.3333"
+    "typescript": "4.1.3"
   },
   "scripts": {
     "update-graphql-types": "dotenv node scripts/update-graphql-types",
@@ -53,7 +54,8 @@
     "build-storybook": "build-storybook -s public",
     "build:styles": "postcss src/tailwind.css -o src/styles.css",
     "prebuild": "yarn build:styles",
-    "prestart": "yarn build:styles"
+    "prestart": "yarn build:styles",
+    "gql:codegen": "graphql-codegen --config codegen.yml"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -66,6 +68,11 @@
   ],
   "devDependencies": {
     "@babel/core": "^7.4.0",
+    "@graphql-codegen/cli": "1.20.0",
+    "@graphql-codegen/fragment-matcher": "^2.0.1",
+    "@graphql-codegen/typescript": "1.20.0",
+    "@graphql-codegen/typescript-operations": "1.17.13",
+    "@graphql-codegen/typescript-react-apollo": "2.2.1",
     "@storybook/addon-actions": "^5.3.18",
     "@storybook/addon-links": "^5.3.18",
     "@storybook/addons": "^5.3.18",
