@@ -1,17 +1,8 @@
-import ApolloClient from "apollo-client";
-import { createHttpLink } from "apollo-link-http";
-import { setContext } from "apollo-link-context";
-import { InMemoryCache } from "apollo-cache-inmemory";
-
-import { IntrospectionFragmentMatcher } from "apollo-cache-inmemory";
-import introspectionQueryResultData from "./fragmentTypes.json";
-
-const fragmentMatcher = new IntrospectionFragmentMatcher({
-  introspectionQueryResultData
-});
+import { ApolloClient, createHttpLink, InMemoryCache } from "@apollo/client";
+import { setContext } from "@apollo/client/link/context";
 
 const httpLink = createHttpLink({
-  uri: process.env.REACT_APP_API_HOST + "/graphql"
+  uri: process.env.REACT_APP_API_HOST + "/graphql",
 });
 
 const authLink = setContext((_, { headers }) => {
@@ -19,12 +10,12 @@ const authLink = setContext((_, { headers }) => {
   return {
     headers: {
       ...headers,
-      Authorization: token ? `Bearer ${token}` : ""
-    }
+      Authorization: token ? `Bearer ${token}` : "",
+    },
   };
 });
 
 export default new ApolloClient({
   link: authLink.concat(httpLink),
-  cache: new InMemoryCache({ fragmentMatcher })
+  cache: new InMemoryCache(),
 });
