@@ -1,7 +1,7 @@
 import gql from "graphql-tag";
 
 export default gql`
-  query branchesQuery($bookPermalink: String!) {
+  query branches($bookPermalink: String!) {
     book(permalink: $bookPermalink) {
       ... on Book {
         id
