@@ -16,28 +16,25 @@ interface CommitProps {
   };
 }
 
-class Commit extends React.Component<CommitProps> {
-  render() {
-    const {
-      permalink,
-      commit: { sha, createdAt, branch },
-      latestCommit,
-    } = this.props;
-    let latest;
-    if (sha != latestCommit.sha) {
-      latest = <Link to={`/books/${permalink}`}> Go to latest revision </Link>;
-    } else {
-      latest = "Latest commit";
-    }
-    return (
-      <div className="text-gray-600 mb-4">
-        <small>
-          {branch.name}@{sha.slice(0, 8)} &middot; {moment(createdAt).fromNow()}{" "}
-          &middot; {latest}
-        </small>
-      </div>
-    );
+const Commit: React.FC<CommitProps> = ({
+  permalink,
+  commit: { sha, createdAt, branch },
+  latestCommit,
+}) => {
+  let latest;
+  if (sha != latestCommit.sha) {
+    latest = <Link to={`/books/${permalink}`}> Go to latest revision </Link>;
+  } else {
+    latest = "Latest commit";
   }
-}
+  return (
+    <div className="text-gray-600 mb-4">
+      <small>
+        {branch.name}@{sha.slice(0, 8)} &middot; {moment(createdAt).fromNow()}{" "}
+        &middot; {latest}
+      </small>
+    </div>
+  );
+};
 
 export default Commit;
