@@ -3,7 +3,7 @@ import ElementWithInfoFragment from "../Notes/ElementWithInfoFragment";
 import NoteFragment from "../Notes/NoteFragment";
 
 export default gql`
-  query noteQuery($bookPermalink: String!, $number: Int!) {
+  query note($bookPermalink: String!, $number: Int!) {
     note(bookPermalink: $bookPermalink, number: $number) {
       ...note
 
