@@ -1,6 +1,6 @@
 import React from "react";
 
-export default () => {
+const PermissionDenied = () => {
   return (
     <div
       className={`bg-white p-4 border-gray-400 border rounded md:w-1/2 text-red-600`}
@@ -9,3 +9,5 @@ export default () => {
     </div>
   );
 };
+
+export default PermissionDenied;
