@@ -7,7 +7,7 @@ import * as serviceWorker from "./serviceWorker";
 import Bugsnag from "@bugsnag/js";
 import BugsnagPluginReact from "@bugsnag/plugin-react";
 
-if (process.env.NODE_ENV != "development") {
+if (process.env.NODE_ENV !== "development") {
   Bugsnag.start({
     apiKey: "5904ec6fda1f66e31d52ebabca2e4fcb",
     plugins: [new BugsnagPluginReact(React)],
