@@ -1,11 +1,11 @@
 import React from "react";
 
-export default (props: any) => {
+const layout = (props: any) => {
   return (
     <div className="bg-gray-200">
       <div className="my-4 mx-auto px-4">
         <menu>
-          <a href="#">
+          <a href="/">
             <strong>Twist</strong>
           </a>
           &nbsp; | &nbsp; Signed in as radar
@@ -15,3 +15,5 @@ export default (props: any) => {
     </div>
   );
 };
+
+export default layout;
